# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/candidates": {
      "post": {
        "operationId": "CandidatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Normalized email or phone already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a normalized candidate record",
        "tags": [
          "Candidates"
        ]
      },
      "get": {
        "operationId": "CandidatesController_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeArchived",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "lastName",
                "email"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": true,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Candidate"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search and filter active candidates",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}": {
      "get": {
        "operationId": "CandidatesController_one",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Candidate not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an active candidate",
        "tags": [
          "Candidates"
        ]
      },
      "patch": {
        "operationId": "CandidatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update candidate details, tags, skills, source, ownership, or consent",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/archive": {
      "post": {
        "operationId": "CandidatesController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a candidate",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/restore": {
      "post": {
        "operationId": "CandidatesController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore an archived candidate",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/duplicates": {
      "get": {
        "operationId": "CandidatesController_duplicates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candidate"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find candidates sharing normalized email or phone",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/merge": {
      "post": {
        "operationId": "CandidatesController_merge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeCandidatesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Both candidates applied for the same job"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Merge a source candidate into this target",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/privacy/export": {
      "get": {
        "operationId": "CandidatesController_export",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "candidate"
                  ],
                  "properties": {
                    "candidate": {
                      "$ref": "#/components/schemas/Candidate"
                    },
                    "applications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Application"
                      }
                    },
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "interviews": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Interview"
                      }
                    },
                    "responses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export candidate personal and recruitment data",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/privacy/deletion-request": {
      "post": {
        "operationId": "CandidatesController_requestDeletion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record an audited personal-data deletion request",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/privacy/erase": {
      "post": {
        "operationId": "CandidatesController_erase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Anonymize candidate personal data while preserving recruitment history",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/candidates/{id}/state": {
      "patch": {
        "operationId": "CandidatesController_state",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change the legacy candidate lifecycle state",
        "tags": [
          "Candidates"
        ]
      }
    },
...
  },
  "info": {
    "title": "GEDPro API",
    "description": "Applicant tracking API for users, candidates, jobs, applications, forms, documents, and interviews.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/v1",
      "description": "Version 1 (recommended)"
    },
    {
      "url": "/",
      "description": "Legacy unversioned compatibility"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "title should not be empty"
            ]
          },
          "path": {
            "type": "string",
            "example": "/jobs"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-15T10:00:00.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "req_01JABCDEF"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCandidateDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "maxLength": 40
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string"
          },
          "ownerId": {
            "type": "number"
          },
          "privacyConsent": {
            "type": "boolean"
          },
          "retentionUntil": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email"
        ]
      },
...
      "Candidate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "normalizedEmail": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "normalizedPhone": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "archivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "privacyConsent": {
            "type": "boolean"
          },
          "consentAt": {
            "format": "date-time",
            "type": "string"
          },
          "retentionUntil": {
            "format": "date-time",
            "type": "string"
          },
          "deletionRequestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "mergedInto": {
            "$ref": "#/components/schemas/Candidate"
          },
          "currentState": {
            "enum": [
              "new",
              "preselected",
              "interview_scheduled",
              "in_interview",
              "accepted",
              "rejected"
            ],
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "normalizedEmail",
          "phone",
          "tags",
          "skills",
          "privacyConsent",
          "currentState",
          "history",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "UpdateCandidateDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "maxLength": 40
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string"
          },
          "ownerId": {
            "type": "number"
          },
          "privacyConsent": {
            "type": "boolean"
          },
          "retentionUntil": {
            "type": "string"
          }
        }
      },
      "MergeCandidatesDto": {
        "type": "object",
        "properties": {
          "sourceCandidateId": {
            "type": "number"
          }
        },
        "required": [
          "sourceCandidateId"
        ]
      },
      "UpdateCandidateStateDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "new",
              "preselected",
              "interview_scheduled",
              "in_interview",
              "accepted",
              "rejected"
            ]
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "state"
        ]
      },
...
      "Application": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "currentStage": {
            "$ref": "#/components/schemas/PipelineStage"
          },
          "source": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "candidate",
          "job",
          "currentStage",
          "history",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
    }
  }
}